+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 ## v0.2
 1. Improved the Marking model to include the Model Extension concern
 2. Improved the EAN-13 Barcode model to include the Model Extension concern
+
+## v0.3
+1. EAN-8 Barcode model & migration
+2. Improved the Ruby Gem Specification to depend on unidom-common v0.4.